# ssr_immemory/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/paint/create-paint": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новую запись о краске",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paints"
                ],
                "summary": "Создать краску",
                "parameters": [
                    {
                        "description": "Данные краски",
                        "name": "paint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paint/{id}": {
            "get": {
                "description": "Возвращает информацию о конкретной краске",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paints"
                ],
                "summary": "Получить краску по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID краски",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paint/{id}/add-to": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет краску в текущую заявку-черновик пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paints"
                ],
                "summary": "Добавить краску в заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID краски",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для добавления",
                        "name": "paintData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddPaintToRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintRequestJSON"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintRequestJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paint/{id}/change-paint": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет данные о краске",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paints"
                ],
                "summary": "Изменить краску",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID краски",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные краски",
                        "name": "paint",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apitypes.PaintJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paint/{id}/delete-paint": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Помечает краску как удаленную",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paints"
                ],
                "summary": "Удалить краску",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID краски",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paint/{id}/upload-image": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Загружает изображение для конкретной краски",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paints"
                ],
                "summary": "Загрузить изображение краски",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID краски",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение краски",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/paints": {
            "get": {
                "description": "Возвращает список всех красок или фильтрует по названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "paints"
                ],
                "summary": "Получить список красок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по названию",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apitypes.PaintJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apitypes.PaintJSON": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "hiding_power": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "boolean"
                },
                "photo": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "apitypes.PaintRequestJSON": {
            "type": "object",
            "properties": {
                "creator_login": {
                    "type": "string"
                },
                "dateFinish": {
                    "type": "string"
                },
                "dateForm": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "min_layers": {
                    "type": "integer"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handler.AddPaintToRequestData": {
            "type": "object",
            "required": [
                "area",
                "layers"
            ],
            "properties": {
                "area": {
                    "type": "number"
                },
                "layers": {
                    "type": "integer"
                }
            }
        }
    }
}